Splits testInProcess.js into tests for separate commands.
